Refractoring general -> Protection de tous les boutons menant à un communication externe contre le multi-click + ajustement de la porté des variable

+++ src/app/item/list-item/shared/del-modal/del-modal.component.ts
@@ -16,6 +16,7 @@ export class DelModalComponent implements OnInit {
   @Input() whoDelete: CategoryComponent | ItemComponent | SubItemComponent;
 
   private deleteForm: FormGroup;
+  private disabledButton;
 
   constructor(private itemService: ItemService,
               private formBuilder: FormBuilder,
@@ -26,34 +27,38 @@ export class DelModalComponent implements OnInit {
   }
 
   initDeleteForm() {
+    this.disabledButton = false;
     this.deleteForm = this.formBuilder.group({
       checkbox: [false, [Validators.pattern('true')]]
     });
   }
   onSubmitDeleteForm() {
-    if ( this.whoDelete instanceof ItemComponent ) {
-      this.itemService.deleteItem(this.whoDelete).catch(
-          reason => {
-            console.log(reason);
+    if (!this.disabledButton) {
+        this.disabledButton = true;
+        if (this.whoDelete instanceof ItemComponent) {
+            this.itemService.deleteItem(this.whoDelete).catch(
+                reason => {
+                    console.log(reason);
+                    this.router.navigate(['/erreur']);
+                }
+            );
+        } else if (this.whoDelete instanceof CategoryComponent) {
+            this.itemService.deleteCategory(this.whoDelete).catch(
+                reason => {
+                    console.log(reason);
+                    this.router.navigate(['/erreur']);
+                }
+            );
+        } else if (this.whoDelete instanceof SubItemComponent) {
+            this.itemService.deleteSubItem(this.whoDelete).catch(
+                reason => {
+                    console.log(reason);
+                    this.router.navigate(['/erreur']);
+                }
+            );
+        } else {
             this.router.navigate(['/erreur']);
-          }
-      );
-    } else if ( this.whoDelete instanceof CategoryComponent ) {
-      this.itemService.deleteCategory(this.whoDelete).catch(
-          reason => {
-            console.log(reason);
-            this.router.navigate(['/erreur']);
-          }
-      );
-    } else if ( this.whoDelete instanceof SubItemComponent ) {
-      this.itemService.deleteSubItem(this.whoDelete).catch(
-          reason => {
-            console.log(reason);
-            this.router.navigate(['/erreur']);
-          }
-      );
-    } else {
-      this.router.navigate(['/erreur']);
+        }
     }
   }
